multi_tenant_file_upload/backend/requirements.txt
django
djangorestframework
gunicorn
psycopg2-binary
dateparser
django-model-utils
pylint
django-session-timeout
django-cors-headers
pytest
pytest-cov
pytest-django-queries
pytest-django
dateparser
requests
boto3
django-storages
Pillow
pycodestyle
flake8
black
ciso8601
djangorestframework-simplejwt
django-bulk-sync
